fix gctp arguments to be int, not long

+++ src/lndsr/space.h
@@ -140,9 +140,9 @@ typedef struct {
 
 typedef struct {
   Space_def_t def;       /* Space definition structure */
-  long (*for_trans)(double lat, double lon, double *x, double *y);
+  int (*for_trans)(double lat, double lon, double *x, double *y);
                          /* Forward transformation function call */
-  long (*inv_trans)(double x, double y, double *lat, double *lon);
+  int (*inv_trans)(double x, double y, double *lat, double *lon);
                          /* Inverse transformation function call */
   double cos_orien;      /* Cosine of the orientation angle */
   double sin_orien;      /* Sine of the orientation angle */
